+++ deep-learning/src/main/python/horovod_installation.sh
