+++ deep-learning/src/main/python/synapse/ml/dl/LitDeepVisionModel.py
