+++ deep-learning/src/main/python/synapse/ml/dl/__init__.py
@@ -2,6 +2,5 @@
 # Licensed under the MIT License. See LICENSE in project root for information.
 
 from synapse.ml.dl.DeepVisionClassifier import *
-from synapse.ml.dl.LightningEstimator import *
-from synapse.ml.dl.LightningParams import *
+from synapse.ml.dl.DeepVisionModel import *
 from synapse.ml.dl.LitDeepVisionModel import *
