+++ deep-learning/src/main/python/tests/dl/conftest.py
@@ -13,7 +13,9 @@ import torchvision.transforms as transforms
 
 IS_WINDOWS = os.name == "nt"
 delimiter = "\\" if IS_WINDOWS else "/"
-dataset_dir = delimiter.join([os.getcwd(), "..", "target"]) + delimiter
+dataset_dir = (
+    delimiter.join([os.getcwd(), os.pardir, os.pardir, os.pardir, "target"]) + delimiter
+)
 
 
 def _download_dataset():
