+++ deep-learning/src/main/python/version.py
